requirements.txt
rank-bm25
faiss-cpu
sentence-transformers
transformers
openai
pydantic
PyMuPDF
tqdm
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
streamlit
pandas
openpyxl
tabulate
docling
pyyaml
easyocr
opencv-python-headless
